Address comments and update README

+++ internal/api/snapshot_backup/snapshot_backup.go
@@ -29,7 +29,7 @@ type SnapshotBackup struct {
 	ID                string            `json:"id"`
 	Progress          Progress          `json:"progress"`
 	ProjectID         string            `json:"projectId"`
-	Retention         int               `json:"retention"`
+	Retention         int64             `json:"retention"`
 	CrossRegionCopies []CrossRegionCopy `json:"crossRegionCopies"`
 	CMEK              CMEKs             `json:"cmek"`
 	Server            Server            `json:"server"`
@@ -39,7 +39,7 @@ type SnapshotBackup struct {
 }
 
 type CreateSnapshotBackupRequest struct {
-	Retention     int      `json:"retention"`
+	Retention     int64    `json:"retention"`
 	RegionsToCopy []string `json:"regionsToCopy"`
 }
 
@@ -52,5 +52,5 @@ type ListSnapshotBackupsResponse struct {
 }
 
 type EditBackupRetentionRequest struct {
-	Retention int `json:"retention"`
+	Retention int64 `json:"retention"`
 }
